requirements.txt
python-telegram-bot[job-queue]==20.7
motor==3.3.2
pymongo==4.5.0
